Refactored the solution set statistics class to support statistics collection based on a vectored iterator pattern.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6292 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-rdf/src/java/com/bigdata/rdf/sparql/ast/EmptySolutionSetStats.java
@@ -51,7 +51,7 @@ public class EmptySolutionSetStats implements ISolutionSetStats {
     }
     
     @Override
-    public int getSolutionSetSize() {
+    public long getSolutionSetSize() {
         return 1;
     }
